+++ src/logoWork.ml
@@ -33,6 +33,39 @@ let make env varname value =
 let name env value varname =
   set_var env varname value
 
+(* TODO 'local' now accepts the form
+   (local varlist varname2 varname3 ...)
+   which should be marked as an error. *)
+let local env varname rest =
+  let create_var env w =
+    try
+      create_var env w
+    with
+    | Failure "create_var" ->
+      error "'local' used outside of any procedure"
+  in
+  begin match varname with
+  | `L w ->
+    create_var env w
+  | `R wl ->
+    List.iter (create_var env) wl
+  end;
+  List.iter (create_var env) rest
+
+let localmake env varname value =
+  let create_var env w =
+    try
+      create_var env w
+    with
+    | Failure "create_var" ->
+      error "'localmake' used outside of any procedure"
+  in
+  create_var env varname;
+  set_var env varname value
+
+let thing env varname =
+  get_var env varname
+
 (** 7.4 Workspace Predicates *)
 
 let definedp env name =
@@ -44,6 +77,9 @@ let namep env name =
 let init env =
   set_pf env "make" Lga.(env @@ word @-> any @-> ret retvoid) make;
   set_pf env "name" Lga.(env @@ any @-> word @-> ret retvoid) name;
+  set_pf env "local" Lga.(env @@ alt word (list word) @-> rest word retvoid) local;
+  set_pf env "localmake" Lga.(env @@ word @-> any @-> ret retvoid) localmake;
+  set_pf env "thing" Lga.(env @@ word @-> ret (value any)) thing;
 
   set_pf env "definedp" Lga.(env @@ word @-> ret (value any)) definedp;
   set_pf env "defined?" Lga.(env @@ word @-> ret (value any)) definedp;