+++ CHANGES.md
@@ -1,3 +1,8 @@
+## 0.2.0 (current)
+
+* Report undefined variables (both global and unassigned locals)
+* Add primitives 'local', 'localmake', 'thing'
+
 ## 0.1.0 (2014-07-26)
 
 * Initial public release